requirements.txt
asgiref==3.8.1
boto3==1.34.139
botocore==1.34.139
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-storages==1.14.3
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jmespath==1.0.1
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
